ron_skills/sql_query_optimizer/api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
anthropic>=0.43.0
pydantic>=2.10.0
python-dotenv>=1.0.0
stripe>=11.0.0
